requirements.txt
--extra-index-url https://test.pypi.org/simple/
hive-metastore-client==1.0.7
python-dateutil==2.8.1
flask==1.1.2
flask-compress==1.9.0
gunicorn===20.0.4
odd-contract-dev==0.0.29
more-itertools==8.6.0
lark-parser==0.11.1
apscheduler==3.7.0
pytz==2020.1